arrange the main function and add option on make file for apply firewall
